Fixed pvp magic animation logic and added websocket setup
Now pvp magic allows the magic to push in and be pushed back based on input, we now have another mage on the other side and speed of  the blue magic is determined by client 1.
Websocket setup has been implemented.
To do: listen to the websocket and send data to update both clients